requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
google==3.0.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
